Kill Message Rewrite
Rewrote the messaging process. Stored messages in an array in the
messages.lua and made a loop in main that shuffles the list. Look over
it and see if you can understand the logic, if you have questions let me
know.

If you have time, test it just to make sure I did it correctly. I tested
it and it seemed fine, but I might've broken something else in the
process.